Part 6 almost done.
levels need to be decorated.

GameFlow need to be applyed.

javadocs like hell....

+++ src/interfaces/LevelInformation.java
@@ -36,7 +36,7 @@ public interface LevelInformation {
      *
      * @return paddle's width in level.
      */
-    int paddleWidth();
+    double paddleWidth();
 
     /**
      * returns the level name to be displayed at the top of the screen.
